# openapi.yaml
openapi: "3.0.1"
info:
  title: "API Documentation"
  version: "1.0"
paths:
  /api/v1/auth/login/userpass:
    post:
      summary: "用户名密码登录"
      description: "用户名密码登录"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "password"
              - "username"
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      accessToken:
                        type: "string"
                        description: ""
                      refreshToken:
                        type: "string"
                        description: ""
                      username:
                        type: "string"
                        description: ""
                      realName:
                        type: "string"
                        description: ""
                      role:
                        type: "string"
                        description: ""
                      branchId:
                        type: "integer"
                        description: ""
                        format: "int64"
                      branchName:
                        type: "string"
                        description: ""
                    description: ""
  /api/v1/auth/login/email:
    post:
      summary: "邮箱验证码登录"
      description: "邮箱验证码登录"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "email"
              - "verifyCode"
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  pattern: "^$|^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      accessToken:
                        type: "string"
                        description: ""
                      refreshToken:
                        type: "string"
                        description: ""
                      username:
                        type: "string"
                        description: ""
                      realName:
                        type: "string"
                        description: ""
                      role:
                        type: "string"
                        description: ""
                      branchId:
                        type: "integer"
                        description: ""
                        format: "int64"
                      branchName:
                        type: "string"
                        description: ""
                    description: ""
  /api/v1/auth/login/phone:
    post:
      summary: "手机号验证码登录"
      description: "手机号验证码登录"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "phone"
              - "verifyCode"
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  pattern: "^$|^1\\d{10}$"
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      accessToken:
                        type: "string"
                        description: ""
                      refreshToken:
                        type: "string"
                        description: ""
                      username:
                        type: "string"
                        description: ""
                      realName:
                        type: "string"
                        description: ""
                      role:
                        type: "string"
                        description: ""
                      branchId:
                        type: "integer"
                        description: ""
                        format: "int64"
                      branchName:
                        type: "string"
                        description: ""
                    description: ""
  /api/v1/auth/verifyCode:
    post:
      summary: "发送验证码"
      description: "发送验证码"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  pattern: "^$|^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  pattern: "^$|^1\\d{10}$"
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "string"
                    description: ""
  /api/v1/auth/logout:
    post:
      summary: "退出登录"
      description: "退出登录"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "refreshToken"
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "string"
                    description: ""
  /api/v1/auth/refresh:
    post:
      summary: "刷新 JWT token"
      description: "刷新 JWT token"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "refreshToken"
              type: "object"
              properties:
                username:
                  type: "string"
                  description: "用户名（用户名密码登录）"
                password:
                  type: "string"
                  description: "密码（用户名密码登录）"
                email:
                  type: "string"
                  description: "邮箱（邮箱验证码登录 / 发送验证码）"
                phone:
                  type: "string"
                  description: "手机号（手机号验证码登录 / 发送验证码，短信通道暂未接入）"
                verifyCode:
                  type: "string"
                  description: "验证码（邮箱/手机号验证码登录）"
                refreshToken:
                  type: "string"
                  description: "刷新令牌（刷新 / 退出登录）"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      accessToken:
                        type: "string"
                        description: ""
                      refreshToken:
                        type: "string"
                        description: ""
                      username:
                        type: "string"
                        description: ""
                      realName:
                        type: "string"
                        description: ""
                      role:
                        type: "string"
                        description: ""
                      branchId:
                        type: "integer"
                        description: ""
                        format: "int64"
                      branchName:
                        type: "string"
                        description: ""
                    description: ""
...